# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Awahan API Documentation",
        "description": "Swagger OpenApi description",
        "contact": {
            "email": "subham.5ine@gmail.com"
        },
        "license": {
            "name": "Nginx 1.22.1",
            "url": "http://nginx.org/LICENSE"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Awahan API Server"
        }
    ],
    "paths": {
...
        "/api/auth/profile-update": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update Profile of Authenticated User",
                "description": "Returns registered user data",
                "operationId": "aa5c9f497a3630931133af2e06dd7a70",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "User First Name",
                                        "type": "string",
                                        "example": "subham"
                                    },
                                    "last_name": {
                                        "description": "User Last Name",
                                        "type": "string",
                                        "example": "saha"
                                    },
                                    "phone": {
                                        "description": "User Phone",
                                        "type": "string",
                                        "example": "7892156160"
                                    },
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "subham.5ine@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/refresh": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh Token of Authenticated User",
                "description": "Returns registered user data",
                "operationId": "62c926773eae265ad5f5e91d498e995b",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register User",
                "description": "Returns registered user data",
                "operationId": "f0e180cf2ce34bf385b5cf9d3169900d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "User First Name",
                                        "type": "string",
                                        "example": "subham"
                                    },
                                    "last_name": {
                                        "description": "User Last Name",
                                        "type": "string",
                                        "example": "saha"
                                    },
                                    "phone": {
                                        "description": "User Phone",
                                        "type": "string",
                                        "example": "7892156160"
                                    },
                                    "email": {
                                        "description": "User Email",
                                        "type": "string",
                                        "example": "subham.5ine@gmail.com"
                                    },
                                    "password": {
                                        "description": "User Password",
                                        "type": "string",
                                        "example": "subham"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/auth/resent-otp/{id}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Re-send Otp",
                "description": "Returns registered user data",
                "operationId": "826051a29755f2d42ef56f1acddf0b99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
        "/api/auth/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update Password of Unauthenticated User",
                "description": "Returns registered user data",
                "operationId": "5df66b266ad585ca403c80f41005cc7a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "otp": {
                                        "description": "User OTP",
                                        "type": "string",
                                        "example": "1234"
                                    },
                                    "old_password": {
                                        "description": "User Old Password",
                                        "type": "string",
                                        "example": "subham"
                                    },
                                    "password": {
                                        "description": "User New Password",
                                        "type": "string",
                                        "example": "subham"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "scheme": "bearer",
                "in": "header",
                "bearerAuth": "bearerAuth",
                "bearerFormat": "JWT"
            },
            "security": {
                "bearerAuth": []
            }
        }
    }
}